add serial term

+++ src/win32/serialport.c
@@ -36,10 +36,10 @@ int bscl_serialport_list_begin(void **sp_list) {
   *sp_list = reg;
   return 0;
 }
-int bscl_serialport_list_next(void *sp_list, char *name, int len) {
+int bscl_serialport_list_next(void *sp_list, char *name, unsigned long len) {
   bscl_serialport_reg_t *reg = sp_list;
   char value_name[20];
-  int value_len = 20;
+  long value_len = 20;
   int ret = RegEnumValue(reg->key, reg->index, value_name, &value_len, NULL, NULL, (LPBYTE)name, &len);
   if (ret != ERROR_SUCCESS) {
     return BSCL_SERIALPORT_EFAIL;
